Display throughput changes in report
It can be tedious to determine how throughput changes based on timing
changes. The two are obviously directly related, but while it's obvious
a 50% reduction in time per iteration results in a 100% improvement in
throughput it's much more cumbersome to say a 26.3% improvement in
time results in a 35.6% improvement in throughput.

I'm not sure how to best display the throughput data alongside the
timing data. I punted that decision for now and just appended another
line below the timing change with a 'thrpt' prefix. This is definitely
not the best way to show the data.